requirements.txt
lxml==4.6.3
Pillow==8.2.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
PyQt5-stubs==5.15.2.0
python-docx==0.8.10
pywin32==300
xlrd==2.0.1
XlsxWriter==1.4.0
